Add note to docs about using different domains when self-hosting.
Fixes #554

Hosting the front-end and back-end on different top-level domains will prevent auth cookies from setting.  You will still be able to log in and use the app, but you will need to re-authenticate every time you refresh the page.

Works: http://growthbook.example.com and http://growthbook-api.example.com
Breaks: http://growthbook-example.com and http://growthbook-api-example.com

This is because of the `SameSite=Lax` setting for our auth cookies.  We could change it to `None`, but that opens up the possibility of CSRF attacks. We have other CSRF protections in place, but multiple layers of protection are always better.