Separate "publishable" and "secret" API keys
This is a pre-requisite for implementing a full REST api. Existing "api keys" are publishable, meaning they can only do limited readonly actions and are safe to reveal to users.

In order to add more powerful REST api endpoints, we need to support secret keys as well.

- [x] Modified existing ApiKey interface to support secret keys
- [x] Updated API Key settings page with separate sections for publishable and secret keys
- [x] Updated all usage of existing API Keys to work with the new structure

Note: This PR does NOT implement any new REST api endpoints. It just adds secret API keys that we can use in the future when we add new endpoints.