Let users enter the experiment_id when creating a draft experiment
We currently only let users enter the experiment name. We then guess the experiment id based on the name.
This behavior is unintuitive and often produces an incorrect id that is hard to debug.

This PR lets the user explicitly enter the experiment id when creating an experiment. If left empty, it will fall back to the guessing behavior that exists now.

The experiment id field is hidden when importing an experiment from a data source or feature rule.