Stronger link between experiments and features
There is no single source of truth for an experiment.  Feature rules define their own targeting conditions, weights, coverage, namespace, etc..  The Experiment analysis also defines a copy of all of these things and it's really easy for them to get out of sync.

This PR makes the Experiment object the source of truth.  Feature rules will simply store a foreign key to the Experiment object along with a map of variation ids to feature values.

This has a number of benefits:
1. As you make changes to an experiment's lifecycle (e.g. start a new phase, stop a test, increase rollout), any associated features will automatically be updated
2. You can change multiple features as part of a single experiment
3. You'll be able to see the full experiment lifecycle on the feature page instead of just a snapshot of the current settings