Feature Flag Experiment References
Currently, feature flag experiment rules define all of the targeting and assignment parameters inline. Then, an entirely separate Experiment object must be created for analysis.

Both the Experiment and the Feature Rule have copies of variation weights, coverage, etc. and it's easy for them to get out-of-sync.

This PR makes the Experiment the source-of-truth. Adding an experiment rule to a feature will just reference the Experiment id instead of defining settings inline.