Remove initial feature rule
Right now, we prompt users to add an initial rule when creating a feature. This is hard to maintain (need 2 copies of the rule UI).  It also gives new users the wrong impression that there's only a single rule for a feature. We want to train users to use the actual rule UI and think of features as having a list of rules.

This PR will also make #1139 much easier to implement.