Revert default MongoDB db name to "test" unless explicitly set
Before updating to the Node.js MongoDB driver v4.X, the default db name was `test`.  Now it is `admin`.  This change was undocumented and was not caught in our initial dependency update PR.

This PR attempts to revert the default behavior to match the old version. It also tries to encourage new users to explicitly set a db name going forward.

1. Updates documentation and starter templates to explicitly set the db name as `growthbook` for new users
2. If dbname is missing from the connection string, make it default to the old `test` value to maintain backwards compatibility.