SDK - Runtime Groups [WIP]
Dynamic Saved Groups that are evaluated in the SDK at runtime. Adds a new `getGroups: (attributes) => Promise<string[]>` context setting in the javascript SDK that lets you return a list of saved groups the user is part of.  This allows for db lookups, API calls, etc. without syncing (and keeping up-to-date) a list of ids within GrowthBook.