Cancel Queries link now actually cancels queries in BigQuery/Athena
Before, the "cancel" queries link would stop waiting for results and update the GrowthBook UI, but the queries would still be executing in the database in the background.

Now, for BigQuery and Athena, we will send a cancel request to the database as well.

Unfortunately, not all db engines support easily cancelling an in-progress query, which is why we're starting with only two.