Fix bug: Changing defaultValue to empty string hides the "publish" CTA
If the default value of a string feature is changed to an empty string and that's the only change in a draft, it is not detected properly and the "publish" button is hidden.

This PR fixes the bug and makes messaging on the page more explicit.  Now, you will always see the "publish" link and we'll just gray it out and show a tooltip if you don't have permission or there are no changes detected.